Removed many UIView extensions properties. I wasn't sure if they were required.
Plus: Split the Layout.swift file into multiple files.
